uploaded project
uploaded a project that compared an optimal Binary Search Tree to a greedy BST. The project was designed to highlight the use of the optimal BST to minimize the comparisons done when searching for a word (minimizing the expected access cost for searching for a key). The balanced search tree is not including because it just aims to make the tree as shallow as possible and does not effectively optimize for searching.